factor into typescript-transformers package
