+++ packages/studio-base/package.json
@@ -50,6 +50,7 @@
     "@foxglove/schemas": "0.4.0",
     "@foxglove/studio": "workspace:*",
     "@foxglove/three-text": "0.1.0",
+    "@foxglove/typescript-transformers": "workspace:*",
     "@foxglove/ulog": "2.1.2",
     "@foxglove/velodyne-cloud": "1.0.1",
     "@foxglove/wasm-bz2": "0.1.0",
