+++ packages/typescript-transformers/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@foxglove/tsconfig/base",
+  "include": ["./src/**/*"],
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./dist",
+    "lib": ["dom", "es2020"]
+  }
+}
